refact: remove unedded versioning
