+++ src/main/java/br/com/kwikecommerce/api/controller/ProductExists.java
@@ -1,6 +1,6 @@
-package br.com.kwikecommerce.api.controller.validator;
+package br.com.kwikecommerce.api.controller;
 
-import br.com.kwikecommerce.api.company.CompanyRepository;
+import br.com.kwikecommerce.api.product.ProductRepository;
 import lombok.RequiredArgsConstructor;
 
 import javax.validation.Constraint;
@@ -13,25 +13,24 @@ import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
 
-@Constraint(validatedBy = CompanyExists.Validator.class)
+@Constraint(validatedBy = ProductExists.Validator.class)
 @Target({ElementType.FIELD})
 @Retention(RetentionPolicy.RUNTIME)
-public @interface CompanyExists {
-
-    String message() default "{v.company.does-not-exists}";
+public @interface ProductExists {
+    String message() default "{v.product.does-not-exists}";
 
     Class<?>[] groups() default {};
 
     Class<? extends Payload>[] payload() default {};
 
     @RequiredArgsConstructor
-    class Validator implements ConstraintValidator<CompanyExists, Long> {
+    class Validator implements ConstraintValidator<ProductExists, Long> {
 
-        private final CompanyRepository companyRepository;
+        private final ProductRepository productRepository;
 
         @Override
-        public boolean isValid(Long companyId, ConstraintValidatorContext constraintValidatorContext) {
-            return companyRepository.existsById(companyId);
+        public boolean isValid(Long productId, ConstraintValidatorContext context) {
+            return productRepository.existsById(productId);
         }
 
     }
