+++ src/main/java/br/com/kwikecommerce/api/controller/category/CategoryApi.java
@@ -1,7 +1,7 @@
 package br.com.kwikecommerce.api.controller.category;
 
-import br.com.kwikecommerce.api.controller.category.dto.CategoryListingResponse;
 import br.com.kwikecommerce.api.controller.category.dto.CategoryCreationRequestDto;
+import br.com.kwikecommerce.api.controller.category.dto.CategoryListingResponse;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import org.springframework.web.bind.annotation.RequestBody;
@@ -16,11 +16,9 @@ import java.util.List;
 )
 public interface CategoryApi {
 
-    @Tag(name = "Categories")
     @Operation(summary = "Create a new category")
     Long create(@RequestBody @Valid CategoryCreationRequestDto categoryCreationRequestDto);
 
-    @Tag(name = "Categories")
     @Operation(summary = "List all categories")
     List<CategoryListingResponse> listAll();
