+++ src/main/java/br/com/kwikecommerce/api/controller/category/CategoryController.java
@@ -1,9 +1,9 @@
 package br.com.kwikecommerce.api.controller.category;
 
-import br.com.kwikecommerce.api.controller.category.dto.CategoryListingResponse;
-import br.com.kwikecommerce.api.controller.category.dto.CategoryCreationRequestDto;
 import br.com.kwikecommerce.api.category.CategoryMapper;
 import br.com.kwikecommerce.api.category.CategoryService;
+import br.com.kwikecommerce.api.controller.category.dto.CategoryCreationRequestDto;
+import br.com.kwikecommerce.api.controller.category.dto.CategoryListingResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -17,7 +17,7 @@ import java.util.List;
 @Validated
 @RequiredArgsConstructor
 @RestController
-@RequestMapping("/v1/categories")
+@RequestMapping("/categories")
 public class CategoryController implements CategoryApi {
 
     private final CategoryService categoryService;
