+++ src/main/java/br/com/kwikecommerce/api/controller/order/OrderApi.java
@@ -2,11 +2,12 @@ package br.com.kwikecommerce.api.controller.order;
 
 import br.com.kwikecommerce.api.application.pagination.PageRequest;
 import br.com.kwikecommerce.api.application.pagination.PageResponseDto;
+import br.com.kwikecommerce.api.application.pagination.sort.NoSortOption;
+import br.com.kwikecommerce.api.controller.OrderExists;
+import br.com.kwikecommerce.api.controller.order.dto.OrderCreationRequestDto;
 import br.com.kwikecommerce.api.controller.order.dto.OrderFindingByFilterResponse;
 import br.com.kwikecommerce.api.controller.order.dto.OrderFindingByIdResponseDto;
-import br.com.kwikecommerce.api.controller.order.dto.OrderCreationRequestDto;
 import br.com.kwikecommerce.api.controller.order.dto.OrderUpdateRequestDto;
-import br.com.kwikecommerce.api.controller.validator.OrderExists;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 
@@ -20,25 +21,18 @@ import javax.validation.constraints.NotNull;
 )
 public interface OrderApi {
 
-    // TODO jhoel.bagnasco 22/08/2021 | Alterar padrão de nomes para verbos no infinitivo
-    @Tag(name = "Orders")
     @Operation(summary = "Initiates an order with its items")
     Long init(@NotNull Long companyId, @Valid OrderCreationRequestDto orderCreationRequestDto);
 
-    @Tag(name = "Orders")
     @Operation(summary = "Updates an order")
     void update(@OrderExists Long orderId, @Valid OrderUpdateRequestDto orderUpdateRequestDto);
 
-    @Tag(name = "Orders")
     void cancel(Long orderId);
 
-    @Tag(name = "Orders")
     @Operation(summary = "Finds orders valid for given filter")
-    PageResponseDto<OrderFindingByFilterResponse> findByFilter(PageRequest pageRequest);
+    PageResponseDto<OrderFindingByFilterResponse> findByFilter(PageRequest<NoSortOption> pageRequest);
 
-    @Tag(name = "Orders")
     @Operation(summary = "Finds an order by its ID")
     OrderFindingByIdResponseDto findById(@OrderExists Long orderId);
-
 
 }
