+++ src/main/resources/application.yml
@@ -1,67 +1,73 @@
 kwik-ecommerce:
-  db:
-    name: ${DB_NAME:kwik_ecommerce_db}
-    host: ${DB_HOST:localhost}
-    port: ${DB_PORT:3306}
-    user: ${DB_USER:root}
-    pass: ${DB_PASS:toor}
-  server:
-    port: ${APP_PORT:8080}
-    context: /api
-  api-info:
-    title: Kwik-Ecommerce API
-    description: API for the Kwik-Ecommerce website.
-    version: 0.0.1
+    db:
+        name: ${DB_NAME:kwik_ecommerce_db}
+        host: ${DB_HOST:localhost}
+        port: ${DB_PORT:3306}
+        user: ${DB_USER:root}
+        pass: ${DB_PASS:toor}
+    server:
+        port: ${APP_PORT:8080}
+        context: /api
+    api-info:
+        title: Kwik Ecommerce API
+        description: API for Kwik Ecommerce.
+        version: 0.0.1
+    aws:
+        access-key: ${AWS_ACCESS_KEY:AKIAUESTIY2WWWFGESXG}
+        secret-key: ${AWS_SECRET_KEY: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}
 
 logging:
-  level:
-    liquibase: INFO
+    level:
+        liquibase: INFO
 
 server:
-  port: ${kwik-ecommerce.server.port}
-  servlet:
-    context-path: ${kwik-ecommerce.server.context}
+    port: ${kwik-ecommerce.server.port}
+    servlet:
+        context-path: ${kwik-ecommerce.server.context}
 
 aws:
-  access-key: AKIAUESTIY2WWWFGESXG
-  secret-key: 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
-  s3:
-    bucket: kwik-ecommerce-dev
-    region: us-east-1
+    access-key: ${kwik-ecommerce.aws.access-key}
+    secret-key: ${kwik-ecommerce.aws.secret-key}
+    s3:
+        bucket: kwik-ecommerce-dev
+        region: us-east-1
 
 spring:
-  banner.location: classpath:/startup-banner.txt
+    banner.location: classpath:/startup-banner.txt
 
-  profiles:
-    active: @activatedProperties@
+    profiles:
+        active: @activatedProperties@
 
-  application:
-    name: kwik-ecommerce-api
+    application:
+        name: kwik-ecommerce-api
 
-  datasource:
-    plataform: mysql
-    url: jdbc:mysql://${kwik-ecommerce.db.host}:${kwik-ecommerce.db.port}/${kwik-ecommerce.db.name}?useSSL=false&createDatabaseIfNotExist=true&allowPublicKeyRetrieval=true
-    username: ${kwik-ecommerce.db.user}
-    password: ${kwik-ecommerce.db.pass}
+    data.rest:
+        detection-strategy: ANNOTATED
 
-  jpa:
-    show-sql: true
-    format-sql: true
+    datasource:
+        plataform: mysql
+        url: jdbc:mysql://${kwik-ecommerce.db.host}:${kwik-ecommerce.db.port}/${kwik-ecommerce.db.name}?useSSL=false&createDatabaseIfNotExist=true&allowPublicKeyRetrieval=true
+        username: ${kwik-ecommerce.db.user}
+        password: ${kwik-ecommerce.db.pass}
 
-  liquibase:
-    contexts: ${spring.profiles.active}
-    enabled: true
-    change-log: classpath:/liquibase/changelog.yml
-    user: ${spring.datasource.username}
-    password: ${spring.datasource.password}
+    jpa:
+        show-sql: true
+        format-sql: true
+
+    liquibase:
+        contexts: ${spring.profiles.active}
+        enabled: true
+        change-log: classpath:/liquibase/changelog.yml
+        user: ${spring.datasource.username}
+        password: ${spring.datasource.password}
 
 springdoc:
-  swagger-ui:
-    path: /ui
-  api-docs:
-    path: /open-api-3/docs
-  produces: application/json
-  group-configs:
-    - group: v1
-      paths-to-match:
-        - /v1/**/*
+    swagger-ui:
+        path: /ui
+    api-docs:
+        path: /open-api-3/docs
+    produces: application/json
+    group-configs:
+        -   group: default
+            paths-to-match:
+                - /**/*
